+++ release-notes.md
@@ -8,6 +8,7 @@ For players:
 * Simplified log filename.
 * Simplified error messages when a mod can't be loaded.
 * Simple nested mod folders are now recognised by SMAPI (e.g. `ModName-1.0\ModName\manifest.json`).
+* Improved TrainerMod command handling & feedback.
 * Fixed game's debug output being shown in the console for all users.
 * Fixed the game-outdated error not pausing before exit.
 * Fixed installer errors for some players when deleting files.